Firmware: select region-aware artifacts with safe fallback
