+++ Meshtastic/Model/Firmware/FirmwareViewModel.swift
@@ -38,9 +38,11 @@ extension FirmwareViewModel {
 class FirmwareViewModel: ObservableObject {
 	@Published var firmwareFiles: [FirmwareFile] = []
 	let hardware: DeviceHardwareEntity
+	let preferredRegion: RegionCodes
 
-	init(forHardware: DeviceHardwareEntity) {
+	init(forHardware: DeviceHardwareEntity, preferredRegion: RegionCodes = .unset) {
 		self.hardware = forHardware
+		self.preferredRegion = preferredRegion
 		Task {
 			refresh()
 		}
@@ -58,18 +60,28 @@ class FirmwareViewModel: ObservableObject {
 		let descriptor = FetchDescriptor<FirmwareReleaseEntity>()
 		do {
 			let firmwareReleases = try context.fetch(descriptor)
-			for release in firmwareReleases {
-				if let architecture = hardwareArchitecture {
-					for firmwareType in FirmwareFile.validFilenameSuffixes(forArchitecture: architecture) {
-						let firmwareFile = try FirmwareFile(firmware: release, hardware: hardware, type: firmwareType)
+				for release in firmwareReleases {
+					let localeTags = preferredRegion == .unset ? [] : preferredRegion.firmwareLocaleTagCandidates
+					if let architecture = hardwareArchitecture {
+						for firmwareType in FirmwareFile.validFilenameSuffixes(forArchitecture: architecture) {
+							let firmwareFile = try FirmwareFile(
+								firmware: release,
+								hardware: hardware,
+								type: firmwareType,
+								localeTags: localeTags
+							)
+							newFirmwareList[firmwareFile.localUrl.lastPathComponent] = firmwareFile
+						}
+					} else {
+						// Just the default
+						let firmwareFile = try FirmwareFile(
+							firmware: release,
+							hardware: hardware,
+							localeTags: localeTags
+						)
 						newFirmwareList[firmwareFile.localUrl.lastPathComponent] = firmwareFile
 					}
-				} else {
-					// Just the default
-					let firmwareFile = try FirmwareFile(firmware: release, hardware: hardware)
-					newFirmwareList[firmwareFile.localUrl.lastPathComponent] = firmwareFile
 				}
-			}
 		} catch {
 			Logger.services.error("Error fetching firmware releases: \(error)")
 		}
